Rename the packages to SafeguardJava
