+++ src/main/java/com/oneidentity/safeguard/safeguardjava/SafeguardConnection.java
@@ -1,12 +1,12 @@
-package com.oneidentity.safeguard.safeguardclient;
+package com.oneidentity.safeguard.safeguardjava;
 
-import com.oneidentity.safeguard.safeguardclient.authentication.IAuthenticationMechanism;
-import com.oneidentity.safeguard.safeguardclient.data.FullResponse;
-import com.oneidentity.safeguard.safeguardclient.data.Method;
-import com.oneidentity.safeguard.safeguardclient.data.Service;
-import com.oneidentity.safeguard.safeguardclient.exceptions.ObjectDisposedException;
-import com.oneidentity.safeguard.safeguardclient.exceptions.SafeguardForJavaException;
-import com.oneidentity.safeguard.safeguardclient.restclient.RestClient;
+import com.oneidentity.safeguard.safeguardjava.authentication.IAuthenticationMechanism;
+import com.oneidentity.safeguard.safeguardjava.data.FullResponse;
+import com.oneidentity.safeguard.safeguardjava.data.Method;
+import com.oneidentity.safeguard.safeguardjava.data.Service;
+import com.oneidentity.safeguard.safeguardjava.exceptions.ObjectDisposedException;
+import com.oneidentity.safeguard.safeguardjava.exceptions.SafeguardForJavaException;
+import com.oneidentity.safeguard.safeguardjava.restclient.RestClient;
 import java.util.HashMap;
 import java.util.Map;
 import javax.ws.rs.core.Response;
