+++ src/main/java/com/oneidentity/safeguard/safeguardjava/authentication/AuthenticatorBase.java
@@ -1,10 +1,10 @@
-package com.oneidentity.safeguard.safeguardclient.authentication;
+package com.oneidentity.safeguard.safeguardjava.authentication;
 
-import com.oneidentity.safeguard.safeguardclient.StringUtils;
-import com.oneidentity.safeguard.safeguardclient.data.AccessTokenBody;
-import com.oneidentity.safeguard.safeguardclient.exceptions.ObjectDisposedException;
-import com.oneidentity.safeguard.safeguardclient.exceptions.SafeguardForJavaException;
-import com.oneidentity.safeguard.safeguardclient.restclient.RestClient;
+import com.oneidentity.safeguard.safeguardjava.StringUtils;
+import com.oneidentity.safeguard.safeguardjava.data.AccessTokenBody;
+import com.oneidentity.safeguard.safeguardjava.exceptions.ObjectDisposedException;
+import com.oneidentity.safeguard.safeguardjava.exceptions.SafeguardForJavaException;
+import com.oneidentity.safeguard.safeguardjava.restclient.RestClient;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.Map;
