+++ src/main/java/com/oneidentity/safeguard/safeguardjava/data/FullResponse.java
@@ -1,4 +1,4 @@
-package com.oneidentity.safeguard.safeguardclient.data;
+package com.oneidentity.safeguard.safeguardjava.data;
 
 import javax.ws.rs.core.MultivaluedMap;
