+++ src/main/java/com/oneidentity/safeguard/safeguardjava/data/Method.java
@@ -1,4 +1,4 @@
-package com.oneidentity.safeguard.safeguardclient.data;
+package com.oneidentity.safeguard.safeguardjava.data;
 
 /**
  * A limited list of methods supported by the Safeguard API. Not all HTTP methods are supported.
